app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
flake8==7.2.0
frozenlist==1.6.0
gotrue==2.12.0
greenlet==3.2.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.57
langchain-openai==0.3.15
langchain-text-splitters==0.3.8
langfuse==2.60.3
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-checkpoint-postgres==2.0.21
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.40
limits==5.1.0
lxml==5.4.0
marshmallow==3.26.1
mccabe==0.7.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.76.2
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
postgrest==1.0.1
primp==0.15.0
propcache==0.3.1
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg-pool==3.2.6
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
structlog==25.3.0
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
passlib[bcrypt] 
python-jose[cryptography]